# backend/api-first/openapi.yaml
openapi: 3.1.0
info:
  title: KOSMO API
  version: "v1"
  description: >-
    API de KOSMO - plataforma low-code de spec-driven development (SDD) asistida por IA.
    Ciclo completo: Descubrimiento -> Características -> Requisitos -> Modelo -> Código.

    Convenciones:
    - Autenticación PKCE (RFC 7636) con JWT RS256 (bearerAuth).
    - Errores RFC 7807 (application/problem+json) con type URN, trace_id y violations, o OAuthErrorResponse (RFC 6749) / HttpErrorResponse.
    - Fechas ISO-8601 UTC con sufijo Z.
    - Las fases siguen el orden Descubrimiento -> Características -> Requisitos -> Modelo -> Código.
      La consistencia se evalúa exclusivamente hacia la derecha (downstream).
  contact:
    name: Equipo KOSMO
    email: dev@kosmo.app
    url: https://github.com/CesarPantoja1/KOSMO
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local — desarrollo en máquina del programador
  - url: https://api-dev.kosmo.app
    description: Desarrollo — entorno de integración continua
  - url: https://api.kosmo.app
    description: Producción — tráfico real de usuarios
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Registrar nuevo usuario
      description: >-
        Crea una nueva cuenta de usuario en KOSMO. La contrasena se procesa con Argon2id (OWASP 2025).
        Si el email ya existe se devuelve 409 Conflict. Limite de velocidad: 3 peticiones / IP / ventana.
      operationId: auth_register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Confirmacion generica de registro (mitiga enumeracion de usuarios CWE-204)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: Datos de registro invalidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthErrorResponse"
        "429":
          description: Rate limit excedido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthErrorResponse"
  /api/v1/auth/authorize:
    post:
      tags: [auth]
      summary: Iniciar autenticacion PKCE — obtener codigo de autorizacion
      description: >-
        Valida las credenciales del usuario e inicia el flujo PKCE. Emite un authorization_code
        de un solo uso (TTL: 5 min) vinculado al code_challenge enviado. Tras 5 intentos fallidos
        consecutivos se responde con 429 y header Retry-After. Limite: 10 peticiones / IP / ventana.
      operationId: auth_authorize
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorizeRequest"
      responses:
        "201":
          description: Codigo de autorizacion emitido. Usar en POST /token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationCodeResponse"
        "401":
          description: Credenciales invalidas (email no encontrado o contrasena incorrecta)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthErrorResponse"
        "429":
          description: Cuenta bloqueada por intentos fallidos excesivos
          headers:
            Retry-After:
              schema:
                type: string
              description: Segundos restantes de bloqueo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthErrorResponse"
              example:
                error: account_locked
                error_description: "Cuenta bloqueada. Intente de nuevo en 300 segundos."
                seconds_remaining: 300
  /api/v1/auth/token:
    post:
      tags: [auth]
      summary: Intercambiar codigo de autorizacion por par de tokens JWT
      description: >-
        Completa el flujo PKCE: valida el code_verifier contra el code_challenge almacenado,
        consume el authorization_code y emite un par de tokens JWT firmados con RS256 (access TTL 15m, refresh TTL 7d).
        Limite: 5 peticiones / IP / ventana.
      operationId: auth_token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenExchangeRequest"
      responses:
        "200":
          description: Par de tokens emitido exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairResponse"
        "400":
          description: Codigo invalido, expirado, consumido o code_verifier no coincide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthErrorResponse"
        "429":
          description: Rate limit excedido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthErrorResponse"
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Renovar par de tokens usando refresh token
      description: >-
        Implementa Token Rotation: consume el refresh token presentado, lo revoca en Redis y emite un par nuevo.
        Si se detecta reutilizacion, toda la familia de tokens queda revocada. Limite: 30 peticiones / IP / ventana.
      operationId: auth_refresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRefreshRequest"
      responses:
        "200":
          description: Par de tokens renovado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairResponse"
        "401":
          description: Refresh token invalido, expirado, revocado o reutilizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthErrorResponse"
        "429":
          description: Rate limit excedido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthErrorResponse"
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Obtener identidad del usuario autenticado
      description: >-
        Decodifica y verifica el access_token del header Authorization: Bearer.
        Devuelve el subject (ID de usuario) y los scopes de la sesion activa.
      operationId: auth_me
      responses:
        "200":
          description: Identidad verificada del token presentado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrincipalView"
        "401":
          description: Token ausente, mal formado, con firma invalida, expirado o revocado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthErrorResponse"
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: Cerrar sesion y revocar tokens activos
      description: >-
        Revoca el access_token del header Authorization: Bearer y opcionalmente el refresh_token del body.
        Ambos quedan invalidados en Redis. Limite: 20 peticiones / IP / ventana.
      operationId: auth_logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "204":
          description: Sesion cerrada exitosamente. Sin cuerpo de respuesta.
        "401":
          description: Token ausente, mal formado, expirado o ya revocado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthErrorResponse"
        "429":
          description: Rate limit excedido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthErrorResponse"

  # ══════════════════════════════ PROJECTS ══════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT RS256 emitido por /api/v1/auth/token o /api/v1/auth/refresh
...
  schemas:
...
    RegisterRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Nombre completo del usuario
          example: Gianfranco
        email:
          type: string
          format: email
          description: Direccion de correo electronico unica
          example: usuario@ejemplo.com
        password:
          type: string
          minLength: 12
          maxLength: 128
          description: Contrasena en texto plano (se hashea con Argon2id)
          example: "M1ContrasenaSegura!"
      required: [name, email, password]
    AuthorizeRequest:
      type: object
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          description: Email del usuario
          example: usuario@ejemplo.com
        password:
          type: string
          minLength: 1
          maxLength: 128
          description: Contrasena en texto plano
          example: "M1ContrasenaSegura!"
        code_challenge:
          type: string
          minLength: 43
          maxLength: 128
          description: Hash SHA-256 del code_verifier en Base64URL sin padding
          example: E9Melhoa2OwvFrEMTJguCHaoeK1t8URWbuGJSstw-cM
        code_challenge_method:
          type: string
          enum: [S256]
          default: S256
          description: Metodo de transformacion PKCE
          example: S256
        scopes:
          type: array
          items:
            type: string
          default: []
          description: Permisos solicitados para la sesion
          example: ["profile:read", "agent:run"]
      required: [email, password, code_challenge]
    TokenExchangeRequest:
      type: object
      additionalProperties: false
      properties:
        grant_type:
          type: string
          enum: [authorization_code]
          description: Debe ser exactamente 'authorization_code'
          example: authorization_code
        code:
          type: string
          minLength: 1
          maxLength: 256
          description: Codigo de autorizacion opaco devuelto por /authorize
          example: a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2
        code_verifier:
          type: string
          minLength: 43
          maxLength: 128
          description: Secreto original generado por el cliente
          example: dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk
      required: [grant_type, code, code_verifier]
    TokenRefreshRequest:
      type: object
      additionalProperties: false
      properties:
        grant_type:
          type: string
          enum: [refresh_token]
          description: Debe ser exactamente 'refresh_token'
          example: refresh_token
        refresh_token:
          type: string
          minLength: 1
          description: JWT de tipo refresh emitido previamente
      required: [grant_type, refresh_token]
    LogoutRequest:
      type: object
      additionalProperties: false
      properties:
        refresh_token:
          type: string
          nullable: true
          description: JWT de tipo refresh a revocar junto con el access token
      required: []
    RegisterResponse:
      type: object
      description: Confirmacion generica de registro para mitigar enumeracion de usuarios (CWE-204)
      properties:
        email:
          type: string
          format: email
          example: usuario@ejemplo.com
        message:
          type: string
          example: Si el correo no estaba registrado previamente, la cuenta ha sido creada exitosamente.
      required: [email, message]
...
    AuthorizationCodeResponse:
      type: object
      properties:
        authorization_code:
          type: string
          description: Codigo de autorizacion opaco de un solo uso
          example: a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2
        expires_in:
          type: integer
          minimum: 1
          maximum: 600
          description: Segundos hasta expiracion
          example: 300
      required: [authorization_code, expires_in]
...
    TokenPairResponse:
      type: object
      properties:
        access:
          $ref: "#/components/schemas/TokenView"
        refresh:
          $ref: "#/components/schemas/TokenView"
        token_type:
          type: string
          enum: [Bearer]
          default: Bearer
          example: Bearer
      required: [access, refresh, token_type]
...
    OAuthErrorResponse:
      type: object
      description: Respuesta de error compatible con RFC 6749 seccion 5.2
      properties:
        error:
          type: string
          description: Codigo de error maquina-legible
          example: invalid_grant
        error_description:
          type: string
          description: Descripcion legible del error
          example: Credenciales invalidas
        seconds_remaining:
          type: integer
          nullable: true
          description: Segundos restantes de bloqueo en caso de error account_locked o rate limit
          example: 300
      required: [error, error_description]

    # ── Projects ──
...